refactor: DomainModelQualifier の移動
