+++ core/src/main/scala/com/iheart/playSwagger/SwaggerSpecRunner.scala
@@ -4,7 +4,7 @@ import java.nio.file.{Files, Paths, StandardOpenOption}
 
 import scala.util.{Failure, Success, Try}
 
-import com.iheart.playSwagger.generator.SwaggerSpecGenerator
+import com.iheart.playSwagger.generator.{PrefixDomainModelQualifier, SwaggerSpecGenerator}
 import play.api.libs.json.{JsValue, Json}
 
 object SwaggerSpecRunner extends App {
