+++ core/src/test/scala/com/iheart/playSwagger/SwaggerParameterMapperSpec.scala
@@ -1,7 +1,7 @@
 package com.iheart.playSwagger
 import com.iheart.playSwagger.domain.CustomTypeMapping
 import com.iheart.playSwagger.domain.parameter.{CustomSwaggerParameter, GenSwaggerParameter}
-import com.iheart.playSwagger.generator.SwaggerParameterMapper
+import com.iheart.playSwagger.generator.{PrefixDomainModelQualifier, SwaggerParameterMapper}
 import org.specs2.mutable.Specification
 import play.api.libs.json.{JsString, Json}
 import play.routes.compiler.Parameter
